[UPDATE] 1층 해금
1. 장애물 플레이어 충돌 시, 재시작
2. 열쇠 습득 및 2층 문 해금
