--- Assets/LowPolyFoodLite/LowPolyFoodLite_2020.3.0_URP_v1.0.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3294b408d2ae84a4ab6cf9b99f3f1628
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
