--- Assets/lsy/Scripts and Animations/Closet.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ad87785502f96074d8c4ba05b3bec98e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
