--- Assets/lsy/Scripts and Animations/Closet/ClosetClosing.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e0b788da15cf52845935b78a74246255
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
