--- Assets/lsy/Scripts and Animations/Closet/ClosetDoor_Base.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6a53b8c24a752874da87800383dbdf1e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
