--- Assets/lsy/Scripts and Animations/Closet/ClosetOpening.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7ec53f8a7d5488646b4529a5308cc732
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
