--- Assets/lsy/Scripts and Animations/Closet/ClosetopencloseDoor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 469e00e876fb40142af620db2b198643
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
