--- Assets/lsy/Scripts and Animations/OvenOpen.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4052a93064f5b2c4ea9b0d7577e6ac73
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
