--- Assets/lsy/Scripts and Animations/OvenOpen/ClosingOven.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 438a7d2af973c3348921e30c83ab5205
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
