--- Assets/lsy/Scripts and Animations/OvenOpen/OpenOven.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 075eb7678139ac54ba0efbab5c4b1255
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
