--- Assets/lsy/Scripts and Animations/OvenOpen/OvenFlip.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ae6447d9d805dbd47a186d4d4b009eb9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
