--- Assets/lsy/Scripts and Animations/OvenOpen/OvenFlip.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86ae241f95cb0464188693a338d04798
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
