+++ Assets/lsy/Scripts and Animations/key.cs
@@ -1,52 +1,48 @@
-//using System.Collections;
-//using System.Collections.Generic;
-//using UnityEngine;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
 
-//public class key : MonoBehaviour
-//{
-//    public bool take;
-//    public GameObject doorkey;
-//    public Transform Player; 
+public class key : MonoBehaviour
+{
+    public GameObject DoorKey;
+    public bool take;
+    public Transform Player;
+    public GameObject door;
+    private Animator doorAnimator;
 
-//    void Start()
-//    {
-//        take = false;
-//    }
+    void Start()
+    {
+        take = false;
+        doorAnimator = door.GetComponent<Animator>();
+    }
 
-//    void OnMouseOver()
-//    {
+    void OnMouseOver()
+    {
 
-//        if (Player)
-//        {
-//            float dist = Vector3.Distance(Player.position, transform.position);
-//            if (dist < 10)
-//            {
-//                if (take == false)
-//                {
+        if (Player)
+        {
+            float dist = Vector3.Distance(Player.position, transform.position);
+            if (dist < 15)
+            {
+                if (take == false)
+                {
+                    if (Input.GetMouseButtonDown(0))
+                    {
+                        UnlockDoor();
+                        Debug.Log("¿­¼è´ç");
+                    }
+                }
+            }
 
-//                }
-//                else
-//                {
-//                    if (take == true)
-//                    {
-//                        if (Input.GetMouseButtonDown(0))
-//                        {
-//                            taken();
-//                        }
-//                    }
+        }
+    }
 
-//                }
-
-//            }
-//        }
+    void UnlockDoor()
+    {
+        DoorKey.SetActive(false);
+        doorAnimator.enabled = true;
+        take = true;
+    }
 
 
-
-//    }
-
-//    private void taken()
-//    {
-//        Debug.Log("¿­¼è È¹µæ");
-//    }
-
-//}
+}
